app/test_requirements.txt
attrs==23.2.0
blinker==1.7.0
click==8.1.7
coverage==7.4.1
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security-Too==5.3.3
Flask-WTF==1.2.1
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mongoengine==0.27.0
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
py==1.11.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pymongo==4.6.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-flakes==4.0.5
pytest-isort==3.1.0
pytest-mypy==0.10.3
pytest-pycodestyle==2.3.1
tomli==2.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
